Added response when in DEBUG mode and not TESTING...
Also removed `pypy` from Travis due to slow testing times and we are not
using it.
