+++ .travis.yml
@@ -2,7 +2,6 @@ language: python
 
 python:
   - '2.7'
-  - 'pypy'
 
 # Enable new container builds.
 sudo: false
